Sing up page creation sep29of25 (#10)
* Update expo version

* Added SignUP page

Note: I haven't been able to get expo go to work for this product on either of my machines today. If the SignUp page doesn't intigrate automatically i.e., creates problems simply by exisiting in the project, I'm expecting that I'll be able to take care of it soon.

Ethan

* REnamed SignUp and deleted the random duplicate that appeared